extractor/requirements.txt
pendulum==3.0.0
psycopg2-binary==2.9.9
boto3==1.34.162
